Add loop type agnostic interface for checkpointing
