+++ enzyme/Enzyme/MLIR/Implementations/AffineAutoDiffOpInterfaceImpl.cpp
@@ -13,35 +13,25 @@
 
 #include "Dialect/Ops.h"
 #include "Implementations/CoreDialectsAutoDiffImplementations.h"
+#include "Implementations/LoopCheckpointing.h"
 #include "Interfaces/AutoDiffOpInterface.h"
 #include "Interfaces/GradientUtilsReverse.h"
 #include "Passes/RemovalUtils.h"
 #include "Passes/Utils.h"
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
+#include "mlir/Dialect/Affine/Utils.h"
 #include "mlir/Dialect/MemRef/IR/MemRef.h"
+#include "mlir/Dialect/SCF/IR/SCF.h"
 #include "mlir/IR/IntegerSet.h"
 #include "llvm/ADT/ScopeExit.h"
+
+#include <array>
 
 using namespace mlir;
 using namespace mlir::enzyme;
 using namespace mlir::affine;
 
 namespace {
-
-// A loop created while differentiating `oldOp` is still doing that loop's work,
-// so it inherits what was set on it. The checkpointing directives are left
-// behind, since the rewrite has already acted on them. Mirrors the helper of
-// the same name in the SCF implementation.
-static void preserveAttributesButCheckpointing(Operation *newOp,
-                                               Operation *oldOp) {
-  for (auto attr : oldOp->getDiscardableAttrs()) {
-    auto name = attr.getName();
-    if (name != "enzyme.enable_checkpointing" &&
-        name != "enzyme.binomial_checkpointing" &&
-        name != "enzyme.checkpoint_period")
-      newOp->setAttr(name, attr.getValue());
-  }
-}
 
 affine::AffineForOp
 createAffineForWithShadows(Operation *op, OpBuilder &builder,
@@ -78,7 +68,262 @@ affine::AffineIfOp createAffineIfWithShadows(Operation *op, OpBuilder &builder,
 
 struct AffineForOpInterfaceReverse
     : public ReverseAutoDiffOpInterface::ExternalModel<
-          AffineForOpInterfaceReverse, affine::AffineForOp> {
+          AffineForOpInterfaceReverse, affine::AffineForOp>,
+      public LoopCheckpointing<AffineForOpInterfaceReverse,
+                               affine::AffineForOp> {
+  // ---- hooks required by LoopCheckpointing<AffineForOpInterfaceReverse,
+  //      affine::AffineForOp> ----
+
+  static std::optional<int64_t>
+  getConstantNumberOfIterations(affine::AffineForOp forOp) {
+    if (!forOp.hasConstantLowerBound())
+      return std::nullopt;
+    if (!forOp.hasConstantUpperBound())
+      return std::nullopt;
+    return (forOp.getConstantUpperBound() - forOp.getConstantLowerBound()) /
+           forOp.getStepAsInt();
+  }
+
+  static Value materializeLowerBound(OpBuilder &builder, Location loc,
+                                     affine::AffineForOp forOp,
+                                     MGradientUtilsReverse *gutils) {
+    SmallVector<Value> ops = llvm::map_to_vector(
+        forOp.getLowerBoundOperands(),
+        [&](Value v) { return gutils->getNewFromOriginal(v); });
+    return affine::AffineApplyOp::create(builder, loc,
+                                         forOp.getLowerBoundMap(), ops);
+  }
+
+  static Value materializeUpperBound(OpBuilder &builder, Location loc,
+                                     affine::AffineForOp forOp,
+                                     MGradientUtilsReverse *gutils) {
+    SmallVector<Value> ops = llvm::map_to_vector(
+        forOp.getUpperBoundOperands(),
+        [&](Value v) { return gutils->getNewFromOriginal(v); });
+    return affine::AffineApplyOp::create(builder, loc,
+                                         forOp.getUpperBoundMap(), ops);
+  }
+
+  static Value materializeStep(OpBuilder &builder, Location loc,
+                               affine::AffineForOp forOp,
+                               MGradientUtilsReverse *) {
+    return arith::ConstantIndexOp::create(builder, loc, forOp.getStepAsInt());
+  }
+
+  static int64_t getConstantStart(affine::AffineForOp forOp) {
+    return forOp.getConstantLowerBound();
+  }
+
+  static int64_t getConstantStep(affine::AffineForOp forOp) {
+    return forOp.getStepAsInt();
+  }
+
+  static LogicalResult requireSingleResultBounds(affine::AffineForOp forOp) {
+    if (forOp.getLowerBound().getMap().getNumResults() != 1 ||
+        forOp.getUpperBound().getMap().getNumResults() != 1)
+      return forOp->emitError()
+             << "cannot differentiate loop with minmax bounds yet";
+    return success();
+  }
+
+  // The checkpointing scaffold substitutes a generically-computed index
+  // Value for the loop's own induction variable, which is not a valid
+  // affine dimension/symbol (it isn't the IV of an affine.for/parallel, nor
+  // a top-level value). Cloning an affine.load/affine.store that indexes
+  // through it verbatim would fail the verifier ("operand cannot be used as
+  // a dimension id"), so instead expand its index map into plain arithmetic
+  // (valid for any index Value, affine or not) via affine::expandAffineMap
+  // -- the same primitive the upstream -lower-affine pass uses -- and emit
+  // memref.load/memref.store. Every other op clones as-is.
+  static void cloneOp(OpBuilder &builder, Operation &op, IRMapping &mapping) {
+    if (auto loadOp = dyn_cast<affine::AffineLoadOp>(&op)) {
+      SmallVector<Value> operands = llvm::map_to_vector(
+          loadOp.getIndices(),
+          [&](Value v) { return mapping.lookupOrDefault(v); });
+      auto indices = affine::expandAffineMap(builder, loadOp.getLoc(),
+                                             loadOp.getAffineMap(), operands);
+      assert(indices && "failed to expand affine.load's index map");
+      Value memref = mapping.lookupOrDefault(loadOp.getMemref());
+      Operation *newLoad =
+          memref::LoadOp::create(builder, loadOp.getLoc(), memref, *indices);
+      mapping.map(loadOp.getResult(), newLoad->getResult(0));
+      // IRMapping tracks an operation map alongside its value map (populated
+      // automatically by builder.clone, which this bypasses); callers such
+      // as publishClonedStep/publishClonedOps and, later,
+      // gutils->getNewFromOriginal(originalLoadOp) (used by
+      // AffineLoadOpInterfaceReverse::cacheValues) all key off it, so it must
+      // be registered explicitly here.
+      mapping.map(&op, newLoad);
+      return;
+    }
+    if (auto storeOp = dyn_cast<affine::AffineStoreOp>(&op)) {
+      SmallVector<Value> operands = llvm::map_to_vector(
+          storeOp.getIndices(),
+          [&](Value v) { return mapping.lookupOrDefault(v); });
+      auto indices = affine::expandAffineMap(
+          builder, storeOp.getLoc(), storeOp.getAffineMap(), operands);
+      assert(indices && "failed to expand affine.store's index map");
+      Value memref = mapping.lookupOrDefault(storeOp.getMemref());
+      Value value = mapping.lookupOrDefault(storeOp.getValue());
+      Operation *newStore = memref::StoreOp::create(builder, storeOp.getLoc(),
+                                                     value, memref, *indices);
+      mapping.map(&op, newStore);
+      return;
+    }
+    builder.clone(op, mapping);
+  }
+
+  // ---- periodic-scaffold hooks (see LoopCheckpointing.h doc comment) ----
+  // Unlike binomial, periodic's decomposition is entirely compile-time
+  // constant, so its scaffold can be built as genuine affine.for -- keeping
+  // any affine.load/affine.store cloned into it valid without needing
+  // cloneOp's memref-lowering fallback.
+
+  static affine::AffineForOp createConstantScaffoldLoop(OpBuilder &builder,
+                                                        Location loc,
+                                                        int64_t lb, int64_t ub,
+                                                        int64_t step,
+                                                        ValueRange inits) {
+    return affine::AffineForOp::create(builder, loc, lb, ub, step, inits);
+  }
+
+  // Nothing to precompute: the bound below is a compile-time AffineMap
+  // attached directly to the loop op, not a separately-emitted runtime
+  // Value, so there is nothing to hoist relative to the mutable-ref cloning
+  // loop the way scf.for's cmpi+select needs to be.
+  static Value computeForwardSegmentBound(OpBuilder &, Location, Value,
+                                          int64_t, int64_t, int64_t) {
+    return Value();
+  }
+
+  // Bound = min(nInner, numIters - outerIV): the standard affine-tiling idiom
+  // for a boundary tile, expressed as a multi-result upper-bound map (the
+  // "min" keyword documented on affine.for). Provably equivalent to
+  // scf.for's cmpi+select formula in every case: the two only disagree where
+  // neither formula's boundary case is active.
+  static affine::AffineForOp
+  createForwardSegmentLoop(OpBuilder &builder, Location loc, Value outerIV,
+                           Value /*precomputedBound*/, int64_t nInner,
+                           int64_t nOuter, int64_t trailingIters,
+                           ValueRange inits) {
+    int64_t numIters = nInner * nOuter + trailingIters;
+    MLIRContext *ctx = builder.getContext();
+    AffineExpr d0 = builder.getAffineDimExpr(0);
+    AffineExpr nInnerExpr = builder.getAffineConstantExpr(nInner);
+    AffineExpr remainingExpr = builder.getAffineConstantExpr(numIters) - d0;
+    AffineMap ubMap = AffineMap::get(1, 0, {nInnerExpr, remainingExpr}, ctx);
+    return affine::AffineForOp::create(builder, loc, /*lbOperands=*/ValueRange{},
+                                       AffineMap::getConstantMap(0, ctx),
+                                       ValueRange{outerIV}, ubMap,
+                                       /*step=*/1, inits);
+  }
+
+  // Unlike scf.for (a plain sequential reverse-iteration counter), this
+  // steps by nInner, matching the outer *forward* loop's own step exactly
+  // (see getReverseOuterScaffoldBounds below).
+  static std::array<int64_t, 3>
+  getReverseOuterScaffoldBounds(int64_t nInner, int64_t nOuter,
+                               int64_t trailingIters) {
+    int64_t s = nOuter + (trailingIters > 0 ? 1 : 0);
+    return {0, nInner * s, nInner};
+  }
+
+  // Reverse counterpart. `outerIV` (call it j') is the *reverse* outer
+  // loop's own induction variable; per getReverseOuterScaffoldBounds, that
+  // loop steps by nInner, so j' = j * nInner where j = 0 (last forward
+  // segment), 1, 2, .... The forward segment index being replayed is
+  // k = (nOuter + hasTrailing - 1) - j, so its base is
+  // segmentBase = k * nInner = (nOuter + hasTrailing - 1) * nInner - j' (no
+  // division needed, since j' is already j * nInner), and the bound is
+  // min(nInner, numIters - segmentBase), expanded into one affine map of j'.
+  // (scf.for's analogous formula uses k = nOuter - j instead, which omits
+  // the "+ hasTrailing - 1" term -- a pre-existing defect, not reproduced
+  // here since this is new code with no existing behavior to preserve; see
+  // the LoopCheckpointing.h doc comment.)
+  static affine::AffineForOp
+  createReverseSegmentLoop(OpBuilder &builder, Location loc, Value outerIV,
+                           int64_t nInner, int64_t nOuter,
+                           int64_t trailingIters, ValueRange inits) {
+    int64_t numIters = nInner * nOuter + trailingIters;
+    int64_t s = nOuter + (trailingIters > 0 ? 1 : 0);
+    // numIters - segmentBase, segmentBase = (s-1)*nInner - j'
+    //   = numIters - (s-1)*nInner + j' = constOffset + j'
+    int64_t constOffset = numIters - (s - 1) * nInner;
+    MLIRContext *ctx = builder.getContext();
+    AffineExpr d0 = builder.getAffineDimExpr(0);
+    AffineExpr nInnerExpr = builder.getAffineConstantExpr(nInner);
+    AffineExpr remainingExpr = builder.getAffineConstantExpr(constOffset) + d0;
+    AffineMap ubMap = AffineMap::get(1, 0, {nInnerExpr, remainingExpr}, ctx);
+    return affine::AffineForOp::create(builder, loc, /*lbOperands=*/ValueRange{},
+                                       AffineMap::getConstantMap(0, ctx),
+                                       ValueRange{outerIV}, ubMap,
+                                       /*step=*/1, inits);
+  }
+
+  static affine::AffineForOp
+  createLoopWithSameBounds(OpBuilder &builder, Location loc,
+                           affine::AffineForOp templateLoop,
+                           ValueRange inits) {
+    return affine::AffineForOp::create(
+        builder, loc, templateLoop.getLowerBoundOperands(),
+        templateLoop.getLowerBoundMap(), templateLoop.getUpperBoundOperands(),
+        templateLoop.getUpperBoundMap(), templateLoop.getStepAsInt(), inits);
+  }
+
+  // (outerIV + localIV) * step + start, as one affine.apply -- the whole
+  // chain must be built as a single affine expression (not the generic
+  // arith ops scf.for's implementation uses) so the result is itself a valid
+  // affine dimension, usable as an index by any affine.load/affine.store
+  // cloned with it substituted in for forOp's own induction variable.
+  static Value computeForwardSegmentIV(OpBuilder &builder, Location loc,
+                                       affine::AffineForOp forOp,
+                                       Value outerIV, Value localIV) {
+    int64_t step = getConstantStep(forOp);
+    int64_t start = getConstantStart(forOp);
+    MLIRContext *ctx = builder.getContext();
+    AffineExpr d0 = builder.getAffineDimExpr(0), d1 = builder.getAffineDimExpr(1);
+    AffineExpr expr = (d0 + d1) * step + start;
+    AffineMap map = AffineMap::get(2, 0, {expr}, ctx);
+    return affine::AffineApplyOp::create(builder, loc, map,
+                                         ValueRange{outerIV, localIV});
+  }
+
+  // (segmentBase + localIV) * step + start, segmentBase = (nOuter +
+  // hasTrailing - 1) * nInner - outerIV -- see createReverseSegmentLoop for
+  // why this differs from scf.for's formula, and for why `outerIV` (j' =
+  // j * nInner) needs no division here to recover segmentBase = k * nInner.
+  static Value computeReverseSegmentBound(OpBuilder &builder, Location loc,
+                                          Value outerIV, int64_t nInner,
+                                          int64_t nOuter,
+                                          int64_t trailingIters) {
+    return Value();
+  }
+
+  static Value computeReverseSegmentIV(OpBuilder &builder, Location loc,
+                                       affine::AffineForOp forOp, Value outerIV,
+                                       Value localIV, int64_t nInner,
+                                       int64_t nOuter, int64_t trailingIters,
+                                       Value precomputedBound) {
+    int64_t s = nOuter + (trailingIters > 0 ? 1 : 0);
+    int64_t lastSegmentBase = (s - 1) * nInner;
+    int64_t step = getConstantStep(forOp);
+    int64_t start = getConstantStart(forOp);
+    MLIRContext *ctx = builder.getContext();
+    AffineExpr jPrime = builder.getAffineDimExpr(0),
+              localIdx = builder.getAffineDimExpr(1);
+    AffineExpr segmentBase = lastSegmentBase - jPrime;
+    AffineExpr flatIV = segmentBase + localIdx;
+    AffineExpr expr = flatIV * step + start;
+    AffineMap map = AffineMap::get(2, 0, {expr}, ctx);
+    return affine::AffineApplyOp::create(builder, loc, map,
+                                         ValueRange{outerIV, localIV});
+  }
+
+  static void createScaffoldYield(OpBuilder &builder, Location loc,
+                                  ValueRange operands) {
+    affine::AffineYieldOp::create(builder, loc, operands);
+  }
+
   LogicalResult createReverseModeAdjoint(Operation *op, OpBuilder &builder,
                                          MGradientUtilsReverse *gutils,
                                          SmallVector<Value> caches) const {
@@ -101,6 +346,29 @@ struct AffineForOpInterfaceReverse
       operandsActive.push_back(!gutils->isConstantValue(operand) ||
                                !gutils->isConstantValue(result));
     }
+
+    // Checkpointing dispatch happens before touching `caches` at all: the
+    // checkpointed schemes cached a completely different layout (see
+    // LoopCheckpointing::cacheBinomial/cachePeriodic) than the plain-loop
+    // path below, so caches[i] must not be popped as a control operand
+    // unless we already know the plain path is the one that applies.
+    if (needsBinomialCheckpointing(forOp) || needsCheckpointing(forOp)) {
+      SmallVector<Value> incomingGradients;
+      for (auto &&[active, res] :
+           llvm::zip_equal(operandsActive, op->getResults())) {
+        if (active) {
+          incomingGradients.push_back(gutils->diffe(res, builder));
+          if (!gutils->isConstantValue(res))
+            gutils->zeroDiffe(res, builder);
+        }
+      }
+      if (auto r = tryCreateReverseModeAdjoint(forOp, op, builder, gutils,
+                                               caches, operandsActive,
+                                               incomingGradients))
+        return *r;
+    }
+
+    // ---- plain (non-checkpointed) reverse path, unchanged ----
 
     SmallVector<Value> revLBOperands, revUBOperands, incomingGradients;
 
@@ -217,6 +485,9 @@ struct AffineForOpInterfaceReverse
   SmallVector<Value> cacheValues(Operation *op,
                                  MGradientUtilsReverse *gutils) const {
     auto forOp = cast<affine::AffineForOp>(op);
+
+    if (auto r = tryCacheValues(forOp, op, gutils))
+      return *r;
 
     SmallVector<Value> caches;
     OpBuilder cacheBuilder(gutils->getNewFromOriginal(op));
@@ -789,8 +1060,13 @@ public:
 
       Value diff = arith::SubIOp::create(builder, forOp->getLoc(), ub, lb);
       if (forOp.getStepAsInt() != 1) {
-        auto step = arith::ConstantIntOp::create(
-            builder, forOp->getLoc(), diff.getType(), forOp.getStepAsInt());
+        // arith::ConstantIntOp requires an IntegerType; diff's type here is
+        // always index (the result of an affine.apply), which
+        // arith::ConstantOp + IntegerAttr handles uniformly (unlike
+        // ConstantIntOp, which asserts on IndexType).
+        auto step = arith::ConstantOp::create(
+            builder, forOp->getLoc(),
+            IntegerAttr::get(diff.getType(), forOp.getStepAsInt()));
         diff = arith::DivUIOp::create(builder, forOp->getLoc(), diff, step);
       }
       return {IntOrValue(diff)};
@@ -808,8 +1084,12 @@ public:
     }
 
     if (forOp.getStepAsInt() != 1) {
-      auto step = arith::ConstantIntOp::create(
-          builder, forOp->getLoc(), val.getType(), forOp.getStepAsInt());
+      // See the identical note in getDimensionBounds above: val's type here
+      // is always index, which needs arith::ConstantOp + IntegerAttr rather
+      // than arith::ConstantIntOp (IntegerType-only).
+      auto step = arith::ConstantOp::create(
+          builder, forOp->getLoc(),
+          IntegerAttr::get(val.getType(), forOp.getStepAsInt()));
       val = arith::DivUIOp::create(builder, forOp->getLoc(), val, step);
     }
     return {val};
@@ -879,6 +1159,11 @@ public:
 void mlir::enzyme::registerAffineDialectAutoDiffInterface(
     DialectRegistry &registry) {
   registry.addExtension(+[](MLIRContext *context, affine::AffineDialect *) {
+    // Binomial checkpointing's scaffold loops are always plain scf ops,
+    // even when checkpointing an affine.for -- load scf here rather than
+    // relying on whatever pass happens to run this interface to have
+    // declared it as a dependent dialect.
+    context->getOrLoadDialect<scf::SCFDialect>();
     registerInterfaces(context);
     affine::AffineLoadOp::attachInterface<AffineLoadOpInterfaceReverse>(
         *context);
